+++ src/server/socket-server/handlers/game.ts
@@ -1,6 +1,6 @@
 import { isLiteralScoreGame } from "@server/game-logic/rooms";
 import { getPenguinString } from "./join";
-import { GameGuard, GameHandler, } from "../handlers";
+import { GameGuard, GameHandler, } from "./handlers";
 
 export const handleLeaveGame: GameHandler<[number]> = (ctx, score) => {
   const { game, data, penguin, msg, prst } = ctx;
