+++ src/server/socket-server/handlers/handlers.ts
@@ -4,7 +4,7 @@ import { PenguinEnvironment, World } from "@server/socket-server/world/world";
 import { WorldGame } from "@server/socket-server/world/world-game";
 import { WorldPenguin } from "@server/socket-server/world/world-penguin";
 import { WorldRoom } from "@server/socket-server/world/world-room";
-import { PenguinMessenger } from "./messenger";
+import { PenguinMessenger } from "../messenger";
 import { GameData } from "@server/timelines/game-data";
 import { SettingsManager } from "@server/settings";
 import { PenguinRepository } from "@server/database/database";
