+++ src/server/socket-server/handlers/join.ts
@@ -4,7 +4,7 @@ import { World } from '@server/socket-server/world/world';
 import { WorldPenguin } from '@server/socket-server/world/world-penguin';
 import { WorldRoom } from '@server/socket-server/world/world-room';
 import { GameData } from '@server/timelines/game-data';
-import { PenguinMessenger } from '../messenger';
+import { PenguinMessenger } from '../../socket-server/messenger';
 import { getClientPuffleIds } from './puffle';
 import { getFurnitureString, getIglooFromId } from './igloo';
 import { WorldTable } from '@server/socket-server/world/world-table';
@@ -15,7 +15,7 @@ import { STARTER_DECKS } from '@server/game-logic/starter-deck';
 import { CARDS } from '@server/game-logic/cards';
 import { choose } from '@common/utils';
 import { SPY_DRILLS_DATA } from '@server/game-logic/spy-drills';
-import { PenguinHandler, PenguinGuard, RoomHandler, WorldContext } from '../handlers';
+import { PenguinHandler, PenguinGuard, RoomHandler, WorldContext } from './handlers';
 
 
 function unequipPuffle(p: WorldPenguin): void {
